Added the unreleased Terentatek model and texture
The Terentatek was meant to be a new boss mob, replacing the Sith Ghost that appeared in Sith Tombs. While slow and somewhat dumb, it would have been highly resistant to lightsabers and force powers, requiring the attacker to resort to creative means to hurt the creature. It would have also dropped it's hide as an item, which would have been used to craft new armors.