Added SithThrone model and texture
(Note: This a BBModel file, not tabula)

This model is for what would have been a Sith Throne block entity. This item would have been a very rare find in a Sith tomb, replacing the coffin. The player would have been able to sit on this throne by right-clicking it and dismount with left-shift, much like how horses work.